requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
requests==2.32.3
pytest-bdd~=8.1.0